Modifications from running the Python hashing code locally
I recently tried running the Python hashing code on my M2 Mac. This PR
reflects some of the things I ran into during that process, including:

- Updates to documentation (fixing some outdated filenames, and adding
a usage-focused README for the Python hashing code)
- Adding a requirements.txt documented needed dependencies, for use with
Python package managers
- Converting the Python code to be more idiomatic for something intended
to be run as a script (vs imported as a module so that its functions may
be used in other files) by insertion of a __main__ clause.
- Log output filepath on successful completion
- Adding output filepath to .gitignore to to avoid accidental commit